# backend/openapi.json
{
    "components": {
        "schemas": {
            "AIConversation": {
                "additionalProperties": false,
                "properties": {
                    "CreatedAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "FamilyID": {
                        "type": "string"
                    },
                    "ID": {
                        "type": "string"
                    },
                    "Title": {
                        "type": "string"
                    },
                    "UpdatedAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "UserID": {
                        "type": "string"
                    }
                },
                "required": [
                    "ID",
                    "FamilyID",
                    "UserID",
                    "Title",
                    "CreatedAt",
                    "UpdatedAt"
                ],
                "type": "object"
            },
            "AIMessage": {
                "additionalProperties": false,
                "properties": {
                    "Content": {
                        "type": "string"
                    },
                    "ConversationID": {
                        "type": "string"
                    },
                    "CreatedAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "ID": {
                        "type": "string"
                    },
                    "Role": {
                        "type": "string"
                    },
                    "ToolCallID": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "ToolCalls": {}
                },
                "required": [
                    "ID",
                    "ConversationID",
                    "Role",
                    "Content",
                    "ToolCalls",
                    "ToolCallID",
                    "CreatedAt"
                ],
                "type": "object"
            },
...
            "ErrorDetail": {
                "additionalProperties": false,
                "properties": {
                    "location": {
                        "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
                        "type": "string"
                    },
                    "message": {
                        "description": "Error message text",
                        "type": "string"
                    },
                    "value": {
                        "description": "The value at the given location"
                    }
                },
                "type": "object"
            },
            "ErrorModel": {
                "additionalProperties": false,
                "properties": {
                    "$schema": {
                        "description": "A URL to the JSON Schema for this object.",
                        "examples": [
                            "https://example.com/schemas/ErrorModel.json"
                        ],
                        "format": "uri",
                        "readOnly": true,
                        "type": "string"
                    },
                    "detail": {
                        "description": "A human-readable explanation specific to this occurrence of the problem.",
                        "examples": [
                            "Property foo is required but is missing."
                        ],
                        "type": "string"
                    },
                    "errors": {
                        "description": "Optional list of individual error details",
                        "items": {
                            "$ref": "#/components/schemas/ErrorDetail"
                        },
                        "type": [
                            "array",
                            "null"
                        ]
                    },
                    "instance": {
                        "description": "A URI reference that identifies the specific occurrence of the problem.",
                        "examples": [
                            "https://example.com/error-log/abc123"
                        ],
                        "format": "uri",
                        "type": "string"
                    },
                    "status": {
                        "description": "HTTP status code",
                        "examples": [
                            400
                        ],
                        "format": "int64",
                        "type": "integer"
                    },
                    "title": {
                        "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
                        "examples": [
                            "Bad Request"
                        ],
                        "type": "string"
                    },
                    "type": {
                        "default": "about:blank",
                        "description": "A URI reference to human-readable documentation for the error.",
                        "examples": [
                            "https://example.com/errors/example"
                        ],
                        "format": "uri",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ListAIConversationsOutputBody": {
                "additionalProperties": false,
                "properties": {
                    "$schema": {
                        "description": "A URL to the JSON Schema for this object.",
                        "examples": [
                            "https://example.com/schemas/ListAIConversationsOutputBody.json"
                        ],
                        "format": "uri",
                        "readOnly": true,
                        "type": "string"
                    },
                    "conversations": {
                        "items": {
                            "$ref": "#/components/schemas/AIConversation"
                        },
                        "type": [
                            "array",
                            "null"
                        ]
                    }
                },
                "required": [
                    "conversations"
                ],
                "type": "object"
            },
            "ListAIMessagesOutputBody": {
                "additionalProperties": false,
                "properties": {
                    "$schema": {
                        "description": "A URL to the JSON Schema for this object.",
                        "examples": [
                            "https://example.com/schemas/ListAIMessagesOutputBody.json"
                        ],
                        "format": "uri",
                        "readOnly": true,
                        "type": "string"
                    },
                    "messages": {
                        "items": {
                            "$ref": "#/components/schemas/AIMessage"
                        },
                        "type": [
                            "array",
                            "null"
                        ]
                    }
                },
                "required": [
                    "messages"
                ],
                "type": "object"
            },
...
            "SendChatMessageInputBody": {
                "additionalProperties": false,
                "properties": {
                    "$schema": {
                        "description": "A URL to the JSON Schema for this object.",
                        "examples": [
                            "https://example.com/schemas/SendChatMessageInputBody.json"
                        ],
                        "format": "uri",
                        "readOnly": true,
                        "type": "string"
                    },
                    "conversation_id": {
                        "description": "ID da conversa existente ou omita para criar uma nova",
                        "type": "string"
                    },
                    "message": {
                        "description": "Mensagem ou comando em linguagem natural",
                        "minLength": 1,
                        "type": "string"
                    },
                    "stream": {
                        "default": false,
                        "description": "Indica se deve responder s\u00edncrono ou via streaming SSE",
                        "type": "boolean"
                    }
                },
                "required": [
                    "message"
                ],
                "type": "object"
            },
            "SendChatMessageOutputBody": {
                "additionalProperties": false,
                "properties": {
                    "$schema": {
                        "description": "A URL to the JSON Schema for this object.",
                        "examples": [
                            "https://example.com/schemas/SendChatMessageOutputBody.json"
                        ],
                        "format": "uri",
                        "readOnly": true,
                        "type": "string"
                    },
                    "conversation_id": {
                        "type": "string"
                    },
                    "message_id": {
                        "type": "string"
                    },
                    "response": {
                        "type": "string"
                    }
                },
                "required": [
                    "conversation_id",
                    "message_id",
                    "response"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "bearerFormat": "JWT",
                "description": "Token JWT emitido no login ou registro.",
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "API do ecossistema Prumo de Gest\u00e3o Financeira Familiar com IA Ag\u00eantica.",
        "title": "Prumo API",
        "version": "1.0.0"
    },
    "openapi": "3.1.0",
    "paths": {
...
        "/v1/ai/chat": {
            "post": {
                "operationId": "sendAIChatMessage",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendChatMessageInputBody"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SendChatMessageOutputBody"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "default": {
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        },
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Envia uma mensagem para o Copilot de IA financeiro",
                "tags": [
                    "Copilot de IA"
                ]
            }
        },
        "/v1/ai/conversations": {
            "get": {
                "operationId": "listAIConversations",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListAIConversationsOutputBody"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "default": {
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        },
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Lista as conversas recentes do Copilot no n\u00facleo familiar",
                "tags": [
                    "Copilot de IA"
                ]
            }
        },
        "/v1/ai/conversations/{id}/messages": {
            "get": {
                "operationId": "listAIMessages",
                "parameters": [
                    {
                        "description": "ID da conversa",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "description": "ID da conversa",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ListAIMessagesOutputBody"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "default": {
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorModel"
                                }
                            }
                        },
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Lista o hist\u00f3rico de mensagens de uma conversa com o Copilot",
                "tags": [
                    "Copilot de IA"
                ]
            }
        },
...
    }
}